fixed errors

+++ src/Calculator.cpp
@@ -67,7 +67,7 @@ static void Calculator::input_ta_event_handler(lv_event_t* e)
 	}
 	#if ENABLE_MCP_KEYPAD
    	if(code == LV_EVENT_KEY_PRESSED){
-	   uint8_t btnID = (uint8_t)lv_event_get_user_data(e);
+	   int btnID = (int)lv_event_get_user_data(e);
 	   cout << btnID << endl;
 	   if(btnID == x_BUTTON){
 		   lv_textarea_add_text(ta,"x");
